bot/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.5
python-telegram-bot==20.8
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
